Create Oven_procedure
Write an IL program to control of a heating oven. The system is started with a Start button
that seals in the Auto mode. This can be stopped if the Stop button (normally closed) is
pushed. When the Auto goes on initially the TON timer is used to sound the horn for the first
5 seconds to warn that the oven will start, and after that the horn stops and the heating coil
start. When the oven is turned off the fan continues to blow for 15s after.